backend/requirements.txt

flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pyjwt==2.8.0
werkzeug==2.3.7
python-dotenv==1.0.0
